Fixed some implicit narrowing conversions

+++ Tasks/EncoderTasks/WavPackEncoderTask.m
@@ -1,7 +1,5 @@
 /*
- *  $Id$
- *
- *  Copyright (C) 2005 - 2007 Stephen F. Booth <me@sbooth.org>
+ *  Copyright (C) 2005 - 2020 Stephen F. Booth <me@sbooth.org>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -67,35 +65,35 @@
 	// Album title
 	album = [metadata albumTitle];
 	if(nil != album)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"ALBUM"] cStringUsingEncoding:NSASCIIStringEncoding], [album UTF8String], strlen([album UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"ALBUM"] cStringUsingEncoding:NSASCIIStringEncoding], [album UTF8String], (int)strlen([album UTF8String]));
 	
 	// Artist
 	artist = [metadata trackArtist];
 	if(nil == artist)
 		artist = [metadata albumArtist];
 	if(nil != artist)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"ARTIST"] cStringUsingEncoding:NSASCIIStringEncoding], [artist UTF8String], strlen([artist UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"ARTIST"] cStringUsingEncoding:NSASCIIStringEncoding], [artist UTF8String], (int)strlen([artist UTF8String]));
 	
 	// Composer
 	composer = [metadata trackComposer];
 	if(nil == composer)
 		composer = [metadata albumComposer];
 	if(nil != composer)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"COMPOSER"] cStringUsingEncoding:NSASCIIStringEncoding], [composer UTF8String], strlen([composer UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"COMPOSER"] cStringUsingEncoding:NSASCIIStringEncoding], [composer UTF8String], (int)strlen([composer UTF8String]));
 	
 	// Genre
 	genre = [metadata trackGenre];
 	if(nil == genre)
 		genre = [metadata albumGenre];
 	if(nil != genre)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"GENRE"] cStringUsingEncoding:NSASCIIStringEncoding], [genre UTF8String], strlen([genre UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"GENRE"] cStringUsingEncoding:NSASCIIStringEncoding], [genre UTF8String], (int)strlen([genre UTF8String]));
 	
 	// Year
 	year = [metadata trackDate];
 	if(nil == year)
 		year = [metadata albumDate];
 	if(nil != year)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"YEAR"] cStringUsingEncoding:NSASCIIStringEncoding], [year UTF8String], strlen([year UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"YEAR"] cStringUsingEncoding:NSASCIIStringEncoding], [year UTF8String], (int)strlen([year UTF8String]));
 	
 	// Comment
 	comment			= [metadata albumComment];
@@ -105,55 +103,55 @@
 	if([[[[self taskInfo] settings] objectForKey:@"saveSettingsInComment"] boolValue])
 		comment = (nil == comment ? [self encoderSettingsString] : [NSString stringWithFormat:@"%@\n%@", comment, [self encoderSettingsString]]);
 	if(nil != comment)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"COMMENT"] cStringUsingEncoding:NSASCIIStringEncoding], [comment UTF8String], strlen([comment UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"COMMENT"] cStringUsingEncoding:NSASCIIStringEncoding], [comment UTF8String], (int)strlen([comment UTF8String]));
 	
 	// Track title
 	title = [metadata trackTitle];
 	if(nil != title)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"TITLE"] cStringUsingEncoding:NSASCIIStringEncoding], [title UTF8String], strlen([title UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"TITLE"] cStringUsingEncoding:NSASCIIStringEncoding], [title UTF8String], (int)strlen([title UTF8String]));
 	
 	// Track number
 	trackNumber = [metadata trackNumber];
 	if(nil != trackNumber)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"TRACK"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [trackNumber intValue]] UTF8String], strlen([[NSString stringWithFormat:@"%u", [trackNumber intValue]] UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"TRACK"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [trackNumber intValue]] UTF8String], (int)strlen([[NSString stringWithFormat:@"%u", [trackNumber intValue]] UTF8String]));
 	
 	// Track total
 	trackTotal = [metadata trackTotal];
 	if(nil != trackTotal)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"TRACKTOTAL"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [trackTotal intValue]] UTF8String], strlen([[NSString stringWithFormat:@"%u", [trackTotal intValue]] UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"TRACKTOTAL"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [trackTotal intValue]] UTF8String], (int)strlen([[NSString stringWithFormat:@"%u", [trackTotal intValue]] UTF8String]));
 	
 	// Disc number
 	discNumber = [metadata discNumber];
 	if(nil != discNumber)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"DISCNUMBER"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [discNumber intValue]] UTF8String], strlen([[NSString stringWithFormat:@"%u", [discNumber intValue]] UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"DISCNUMBER"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [discNumber intValue]] UTF8String], (int)strlen([[NSString stringWithFormat:@"%u", [discNumber intValue]] UTF8String]));
 	
 	// Discs in set
 	discTotal = [metadata discTotal];
 	if(nil != discTotal)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"DISCTOTAL"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [discTotal intValue]] UTF8String], strlen([[NSString stringWithFormat:@"%u", [discTotal intValue]] UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"DISCTOTAL"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [discTotal intValue]] UTF8String], (int)strlen([[NSString stringWithFormat:@"%u", [discTotal intValue]] UTF8String]));
 	
 	// Compilation
 	compilation = [metadata compilation];
 	if(nil != compilation)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"COMPILATION"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [compilation intValue]] UTF8String], strlen([[NSString stringWithFormat:@"%u", [compilation intValue]] UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"COMPILATION"] cStringUsingEncoding:NSASCIIStringEncoding], [[NSString stringWithFormat:@"%u", [compilation intValue]] UTF8String], (int)strlen([[NSString stringWithFormat:@"%u", [compilation intValue]] UTF8String]));
 	
 	// ISRC
 	isrc = [metadata ISRC];
 	if(nil != isrc)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"ISRC"] cStringUsingEncoding:NSASCIIStringEncoding], [isrc UTF8String], strlen([isrc UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"ISRC"] cStringUsingEncoding:NSASCIIStringEncoding], [isrc UTF8String], (int)strlen([isrc UTF8String]));
 	
 	// MCN
 	mcn = [metadata MCN];
 	if(nil != mcn)
-		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"MCN"] cStringUsingEncoding:NSASCIIStringEncoding], [mcn UTF8String], strlen([mcn UTF8String]));
+		WavpackAppendTagItem(wpc, [[AudioMetadata customizeWavPackTag:@"MCN"] cStringUsingEncoding:NSASCIIStringEncoding], [mcn UTF8String], (int)strlen([mcn UTF8String]));
 	
 	// Encoded by
 	bundleVersion = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleVersion"];
-	WavpackAppendTagItem(wpc, "TOOL NAME", "Max", strlen("Max"));
-	WavpackAppendTagItem(wpc, "TOOL VERSION", [bundleVersion UTF8String], strlen([bundleVersion UTF8String]));
+	WavpackAppendTagItem(wpc, "TOOL NAME", "Max", (int)strlen("Max"));
+	WavpackAppendTagItem(wpc, "TOOL VERSION", [bundleVersion UTF8String], (int)strlen([bundleVersion UTF8String]));
 	
 	// Encoder settings
-	WavpackAppendTagItem(wpc, "ENCODING", [[self encoderSettingsString] UTF8String], strlen([[self encoderSettingsString] UTF8String]));	
+	WavpackAppendTagItem(wpc, "ENCODING", [[self encoderSettingsString] UTF8String], (int)strlen([[self encoderSettingsString] UTF8String]));
 
 	result = WavpackWriteTag(wpc);
 	NSAssert(0 != result, NSLocalizedStringFromTable(@"Unable to write to the output file.", @"Exceptions", @""));
